+++ docs/changes.rst
@@ -8,6 +8,7 @@ Release 2.0.0
  * Fix :mod:`pytiger.logging.config` Python 3.13 compatibility.
  * Remove all CentOS 7 packages, including SCL.
  * Remove deprecated :mod:`pytiger.logging.legacy` module.
+ * Remove deprecated :mod:`pytiger.nagios` module.
 
 Release 1.2.2
 =============