+++ docs/changes.rst
@@ -5,11 +5,12 @@ Changes in pytiger
 Release 2.0.0
 =============
 
- * Fix :mod:`pytiger.logging.config` Python 3.13 compatibility.
+ * Drop Python 2 compatibility.
  * Remove all CentOS 7 packages, including SCL.
  * Remove deprecated :mod:`pytiger.logging.legacy` module.
  * Remove deprecated :mod:`pytiger.nagios` module.
  * Remove deprecated :func:`pytiger.monitoring.MonitoringCheck.warn` method.
+ * Fix :mod:`pytiger.logging.config` Python 3.13 compatibility.
 
 Release 1.2.2
 =============