+++ docs/changes.rst
@@ -5,7 +5,8 @@ Changes in pytiger
 Release 2.0.0
 =============
 
- * Drop Python 2 compatibility.
+ * Drop Python 2 compatibility. The minimum supported Python version is now
+   3.6.
  * Remove all CentOS 7 packages, including SCL.
  * Remove deprecated :mod:`pytiger.logging.legacy` module.
  * Remove deprecated :mod:`pytiger.nagios` module.